Initial version of a program to run other programs in the background, detached from the current terminal, even if the program itself doesn't support it.
run-as-daemon detaches completely from the current terminal, starts a
new session as a group leader, opens all output file descriptors to
point to the logging destination, writes its process ID into a PID
file, and runs the given command.

run-as-daemon optionally also supports logging into a file or logging
into a logger(1) instance running as a child of the process executed.